Exctraction of proto library from base, essentials. This commit breaks the build. The following commits will fix it.
