+++ opencog/atoms/CMakeLists.txt
@@ -1,4 +1,5 @@
 
+ADD_SUBDIRECTORY (proto)
 ADD_SUBDIRECTORY (base)
 ADD_SUBDIRECTORY (core)
 ADD_SUBDIRECTORY (execution)
