+++ opencog/atoms/base/CMakeLists.txt
@@ -1,33 +1,22 @@
 
-OPENCOG_ADD_ATOM_TYPES(atom_types.script 
-  atom_types.h
-  atom_types.definitions
-  atom_types.inheritance
-  core_types.scm
-  core_types.pyx)
-ADD_CUSTOM_TARGET(opencog_atom_types DEPENDS atom_types.h)
-
 # The atom_types.h file is written to the build directory
 INCLUDE_DIRECTORIES(${CMAKE_BINARY_DIR})
 
 ADD_LIBRARY (atombase
-	atom_types_init.cc
 	Atom.cc
 	ClassServer.cc
-	FloatValue.cc
 	Handle.cc
 	Link.cc
-	LinkValue.cc
 	Node.cc
-	StringValue.cc
 	Valuation.cc
 )
 
 # Without this, parallel make will race and crap up the generated files.
-ADD_DEPENDENCIES(atombase opencog_atom_types)
+ADD_DEPENDENCIES(atombase opencog_atom_types atomproto)
 
 TARGET_LINK_LIBRARIES(atombase
-	truthvalue  
+	atomproto
+	truthvalue
 	${COGUTIL_LIBRARY}
 )
 
@@ -37,23 +26,10 @@ INSTALL (TARGETS atombase
 
 INSTALL (FILES
 	Atom.h
-	${CMAKE_CURRENT_BINARY_DIR}/atom_types.h
-	atom_types.cc
 	ClassServer.h
-	FloatValue.h
 	Handle.h
 	Link.h
-	LinkValue.h
 	Node.h
-	ProtoAtom.h
-	StringValue.h
-	types.h
 	Valuation.h
 	DESTINATION "include/opencog/atoms/base"
-)
-
-# Install the auto-generated atom types as well
-INSTALL (FILES
-	${CMAKE_CURRENT_BINARY_DIR}/core_types.scm
-   DESTINATION "${DATADIR}/scm/opencog/base"
 )
