+++ opencog/atoms/proto/atom_types.cc
