+++ opencog/atoms/proto/atom_types.script
