+++ opencog/atoms/proto/types.h
